обновил подключение для cJSON и изменил миграции

+++ migrations.sql
@@ -1,5 +1,5 @@
 create table Perfume (
-    perfume_id bigint primary key,
+    perfume_id serial primary key,
     perfume_name varchar(64) not null,
     brand_id bigint not null references Brands(brand_id),
     category_id bigint not null references Categories(category_id),
@@ -12,14 +12,14 @@ create table Perfume (
 );
 
 create table Categories (
-    category_id bigint primary key,
+    category_id serial primary key,
     Category_name varchar(64) not null,
     updated_at timestamptz,
     created_at timestamptz not null
 );
 
 create table Bottles (
-    bottle_id bigint primary key,
+    bottle_id serial primary key,
     bottle_name varchar(64) not null,
     volume smallint not null,
     material varchar(64) not null,
@@ -28,14 +28,14 @@ create table Bottles (
 );
 
 create table Brands (
-    brand_id bigint primary key,
+    brand_id serial primary key,
     brand_name varchar(64) not null,
     updated_at timestamptz,
     created_at timestamptz not null
 );
 
 create table Composition (
-    composition_id bigint primary key,
+    composition_id serial primary key,
     perfume_id bigint not null references Perfume(perfume_id),
     ingredient_id bigint not null references Ingredient(ingredient_id),
     percent tinyint not null,
@@ -44,14 +44,14 @@ create table Composition (
 );
 
 create table Ingredient (
-    ingredient_id bigint primary key,
+    ingredient_id serial primary key,
     ingredient_name varchar(64) not null,
     updated_at timestamptz,
     created_at timestamptz not null
 );
 
 create table Payment (
-    payment_id bigint primary key,
+    payment_id serial primary key,
     order_id bigint not null references Orders(order_id),
     payment_date date,
     payment_method varchar(64) not null,
@@ -60,7 +60,7 @@ create table Payment (
 );
 
 create table Orders (
-    order_id bigint primary key,
+    order_id serial primary key,
     customer_id bigint not null references Customer(customer_id),
     order_date date,
     total_amount decimal(10,2),
@@ -69,7 +69,7 @@ create table Orders (
 );
 
 create table Delivery (
-    delivery_id bigint primary key,
+    delivery_id serial primary key,
     order_id bigint not null references Orders(order_id),
     delivery_date date,
     address text not null,
@@ -78,7 +78,7 @@ create table Delivery (
 );
 
 create table Customer (
-    customer_id bigint primary key,
+    customer_id serial primary key,
     customer_name varchar(256) not null,
     email varchar(128) not null,
     phone varchar(11) not null,
@@ -87,7 +87,7 @@ create table Customer (
 );
 
 create table Review (
-    review_id bigint primary key,
+    review_id serial primary key,
     perfume_id bigint not null references Perfume(perfume_id),
     customer_id bigint not null references Customer(customer_id),
     rating int,
@@ -97,7 +97,7 @@ create table Review (
 );
 
 create table Supplier (
-    supplier_id bigint primary key,
+    supplier_id serial primary key,
     supplier_name varchar(128) not null,
     contact_person varchar(128) not null,
     phone varchar(11) not null,
@@ -106,7 +106,7 @@ create table Supplier (
 );
 
 create table Supply (
-    supply_id bigint primary key,
+    supply_id serial primary key,
     supplier_id bigint not null references Supplier(supplier_id),
     perfume_id bigint not null references Perfume(perfume_id),
     quantity smallint not null,
@@ -116,7 +116,7 @@ create table Supply (
 );
 
 create table Promotion (
-    promotion_id bigint primary key,
+    promotion_id serial primary key,
     promotion_name varchar(128) not null,
     start_date date,
     end_date date,
@@ -125,7 +125,7 @@ create table Promotion (
 );
 
 create table PromotionParticipation (
-    participation_id bigint primary key,
+    participation_id serial primary key,
     perfume_id bigint not null references Perfume(perfume_id),
     promotion_id bigint not null references Promotion(promotion_id),
     updated_at timestamptz,
@@ -133,14 +133,14 @@ create table PromotionParticipation (
 );
 
 create table Locations (
-    location_id bigint primary key,
+    location_id serial primary key,
     location_name varchar(256) not null,
     updated_at timestamptz,
     created_at timestamptz not null
 );
 
 create table Shelf (
-    shelf_id bigint primary key,
+    shelf_id serial primary key,
     store_id bigint not null references Store(store_id),
     shelf_number int not null,
     updated_at timestamptz,
@@ -148,7 +148,7 @@ create table Shelf (
 );
 
 create table Store (
-    store_id bigint primary key,
+    store_id serial primary key,
     perfume_id bigint not null references Perfume(perfume_id),
     store_name varchar(256) not null,
     location_id bigint not null references Locations(location_id),
@@ -157,7 +157,7 @@ create table Store (
 );
 
 create table Employee (
-    employee_id bigint primary key,
+    employee_id serial primary key,
     employee_name varchar(128) not null,
     store_id bigint not null references Store(store_id),
     position varchar(128) not null,
@@ -166,7 +166,7 @@ create table Employee (
 );
 
 create table Salary (
-    salary_id bigint primary key,
+    salary_id serial primary key,
     employee_id bigint not null references Employee(employee_id),
     month date,
     amount decimal(10, 2),
